backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
groq==0.4.1
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.24.3
python-dotenv==1.0.0
setuptools